# WireFormatParser Performance Optimizations