## Summary

This PR optimizes WireFormatParser for better performance through cache locality improvements and algorithmic optimizations:

- **ThreadLocal parser caching**: Reuse constructed parsers via ThreadLocal cache to avoid repeated descriptor processing
- **Primitive arrays**: Replace object arrays with primitive int/byte arrays for better cache locality (field numbers, wire types)
- **Field mapping inlining**: Eliminate helper methods and inline field mapping construction
- **Optimized null checks**: Use `ne null` instead of `!= null` for better JVM optimization
- **Direct field type access**: Store and access `FieldDescriptor.Type` directly instead of descriptor array lookups
- **Wire type validation**: Add `fieldWireTypes` array for optimized parseFieldWithState validation
- **Specialized repeated parsing**: Split parseRepeatedFieldWithState into specialized packed/unpacked methods
- **skipRawBytes**: Replace readByteArray with skipRawBytes where data isn't needed

## Key Changes

- `WireFormatParser.scala`: Complete refactoring for performance (291 insertions, 284 deletions)
  - ThreadLocal parser construction cache
  - Primitive arrays for field metadata
  - Specialized methods for packed vs unpacked repeated fields
  - Optimized field lookup and validation paths

## Performance Impact

These optimizations improve WireFormatParser throughput through:
- Reduced memory allocations (ThreadLocal caching, primitive arrays)
- Better CPU cache locality (primitive int/byte arrays vs object arrays)
- Reduced method call overhead (inlined field mapping)
- Faster field validation (precomputed wire types)
- Specialized code paths (packed vs unpacked)

## Testing

- Compiles successfully: `sbt --error \"core/compile\"`
- Existing tests pass
- Performance benchmarks verify improvements

## Dependencies

Builds on PR #29 (recursive parsing support).

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>