@bumfo 2025-09-30
## Performance Benchmark Results

Compared PR3 optimizations against baseline (PR2):

### Results

| Benchmark | Baseline (ns/op) | PR3 (ns/op) | Improvement |
|-----------|------------------|-------------|-------------|
| simpleDirectWireFormatParser | 2331.452 ± 9.861 | 2026.589 ± 26.043 | **1.15x speedup** (13% reduction) |
| simpleGeneratedWireFormatParser | 1830.886 ± 30.878 | 1830.382 ± 31.982 | No change |
| simpleProtoParsing | 2385.865 ± 12.102 | 2288.019 ± 10.611 | 1.04x speedup (4% reduction) |

### Analysis

The performance optimizations deliver:
- **13% execution time reduction** for direct wire format parsing (2331ns → 2027ns)
- Generated wire format parser unchanged (already optimal from code generation)
- **4% execution time reduction** for message-based parsing

The primitive arrays, ThreadLocal caching, and inlined field mappings primarily benefit the direct WireFormatParser path, achieving the targeted performance improvements through better cache locality and reduced allocations.